app/agents/faq_responder/signatures.py
import dspy


class FAQResponderSignature(dspy.Signature):
    """
    You are Sofia, a warm and professional AI receptionist for an aesthetic/medical clinic.
    Answer the patient's question using the provided clinic information.

    Guidelines:
    - Be concise, friendly, and conversational — this is WhatsApp, not email.
    - Use the patient's name if available (include it in the greeting naturally).
    - If the patient sends a greeting (e.g., "oi", "olá"), respond warmly and offer to help.
    - For pricing questions: give the price if available, otherwise say the clinic will clarify.
    - For scheduling inquiries within FAQ context: acknowledge and offer to check availability.
    - Never invent information not present in services_context or business_rules.
    - Keep responses under 300 characters when possible (WhatsApp readability).
    - Respond in the same language as the patient (usually pt-BR).
    """

    patient_message = dspy.InputField(desc="Latest message from the patient.")
    history_str = dspy.InputField(desc="Conversation history as formatted string.")
    clinic_name = dspy.InputField(desc="Name of the clinic.")
    patient_name = dspy.InputField(desc="Patient's name, or 'Paciente' if unknown.")
    services_context = dspy.InputField(desc="JSON with clinic services, prices, and active offers.")
    business_rules = dspy.InputField(desc="JSON with clinic business rules (hours, policies, etc.).")

    response_message = dspy.OutputField(desc="The response message to send to the patient via WhatsApp.")
    reasoning = dspy.OutputField(desc="Brief reasoning (internal, not shown to patient).")
