requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2

dspy-ai==2.5.43
langgraph==0.2.60

openai==1.57.2
anthropic==0.40.0
groq==0.12.0

supabase==2.10.0
psycopg[binary]==3.2.13

python-dotenv==1.0.1
python-multipart==0.0.18
httpx==0.27.2
